fix: init minicpm3

+++ src/backends/cpu/op/CPULayerNorm.hpp
@@ -29,7 +29,7 @@ class CPULayerNormCreator : public CPUBackend::Creator {
 public:
     virtual Op *create(OpParam op_param, Backend *bn, string name, int threadCount) const {
         bool bias = (bool)op_param["bias"];
-        int normSize = (int)op_param["norm_size"];
+        float normSize = (float)op_param["norm_size"];
         int epsilon = (int)op_param["epsilon"];
         return new CPULayerNorm(bn, name, normSize, bias, epsilon, threadCount);
     }
